Move `FeedRefreshController` to seperatre file
