+++ EssentialFeed/EssentialFeediOS/Controllers/FeedRefreshController.swift
@@ -1,0 +1,35 @@
+//
+//  FeedRefreshController.swift
+//  EssentialFeediOS
+//
+//  Created by Prabhat Tiwari on 07/07/25.
+//
+
+import UIKit
+import EssentialFeed
+
+final public class FeedRefreshController: NSObject {
+    public lazy var view: UIRefreshControl = {
+        let view = UIRefreshControl()
+        view.addTarget(self, action: #selector(refresh), for: .valueChanged)
+        return view
+    }()
+    
+    private let feedLoader: FeedLoader
+    
+    var onRefresh: (([FeedImage]) -> Void)?
+    
+    init(feedLoader: FeedLoader) {
+        self.feedLoader = feedLoader
+    }
+    
+    @objc func refresh() {
+        view.beginRefreshing()
+        feedLoader.load { [weak self] result in
+            if let feed = try? result.get() {
+                self?.onRefresh?(feed)
+            }
+            self?.view.endRefreshing()
+        }
+    }
+}
