+++ EssentialFeed/EssentialFeediOS/Controllers/FeedViewController.swift
@@ -8,32 +8,6 @@
 import UIKit
 import EssentialFeed
 
-final public class FeedRefreshController: NSObject {
-    public lazy var view: UIRefreshControl = {
-        let view = UIRefreshControl()
-        view.addTarget(self, action: #selector(refresh), for: .valueChanged)
-        return view
-    }()
-    
-    private let feedLoader: FeedLoader
-    
-    var onRefresh: (([FeedImage]) -> Void)?
-    
-    init(feedLoader: FeedLoader) {
-        self.feedLoader = feedLoader
-    }
-    
-    @objc func refresh() {
-        view.beginRefreshing()
-        feedLoader.load { [weak self] result in
-            if let feed = try? result.get() {
-                self?.onRefresh?(feed)
-            }
-            self?.view.endRefreshing()
-        }
-    }
-}
-
 final public class FeedViewController: UITableViewController, UITableViewDataSourcePrefetching {
     
     public var refreshController: FeedRefreshController?
